Adds a bugfix to make tests run on Nexus 5X
This fix adds a tearDown method to all the persistence text, which shuts
down the ContentProvider after each test. This is a try to fix tests
not running on Nexus 5X, because of data duplication because of unclean
database shutdown. This fix is not necessary to run the tests for example
on the Nexus 5, but it also is no problem on those devices, so we keep
it until we know more about the bug/feature.