Fix timeout for isRunning
This commit fixes the wrong call to the Handler#postAtTime method used
to implement the timeout in the PongReceiver. It has been called with a
relative time but needs to be called with an absolute time in
milliseconds since SystemClock#uptimeMillis().

The commit also adds some missing documentation and fixes the service
tests to run at least with the most recent version of the emulator.