Fix #MOV-169
This commit is in preparation of running instrumentation tests in the CI
system. Since the Android emulator supports geo locations and sensors
now this should work as soon as all instrumented tests run again. This
fixes the problem that locations where not deleted properly. It also
adds some documentation and moves the MovebisDataTransmissionTest to
ignore. The latter is necessary since it requires a running Movebis
server, which we can not start locally by now.