Add stop service message to background service
This commit adds a message that is send by the DataCapturingBackgroundService to the DataCapturingService at the end of onDestroy. That message informs the DataCapturingService that the background service has stopped. This should work more reliable then the broadcast used so far. To be on the save side we leave the broadcast there as well, but might remove it in a future update.