Fixes #MOV-233 by not loading mid from db
In contrast to a previous fix, where storeData was fixed to not load a
measurement identifier from the database each time, storeLocation was
not handled the same way. It still has been loading the measurement
identifier from the database. This commit fixes this state by using the
cached measurement identifier.