Makes non running test flaky and ignore
The DataCapturingServiceTestWithoutPermission will only work on the
first invocation. The reason is, that in contrast to the documentation,
Android does not set permissions on each test. If you call the
GrantPermissionRule once, that permission is fixed until removed again,
manually or until the test app has been reinstalled. Until this is
fixed, we need to consider this as a FlakyTest and remove it from the
regular test execution.