Add authority as parameter to other projects
The synchronization and the datacapturing project both required some
adaptations since the persistence context is used by both. This mostly
involves adding the authority as a parameter and make sure all tests use
that parameter.

For Movebis de.cyface.provider is the hard coded authority to keep
downward compatibility.

Problem that still exists is that you need to set the same provider when
creating a DataCapturingService as well as in your manifest. The
manifest also is required to override the default content provider as
declared by the persistence project. This needs to be done by each using
application separately.