Start to fix test. WIP commit - last state before going back to database #CY-4405
We do not merge this back into the dev as we decided that it makes more sense to use the already tested code
for the gps points, the measurement state etc. as the database offers much logic which we would not re-invent.
However, we'll cherry-pick all good changes and use the file persistence code to store the acceleration points.