Simplifies writeSerializedCompressed() MOV-719
We do this as we think that the zombie temp files reported in MOV-719 might have been created by that messy code.
Additionally we saw "stream not closed" warnining which might have been caused by that code, too.
Thus, we deleted the workaround "cleaning" code in the SyncAdapter to see if that is not necesarry.