Fix ShutdownFinishedHandler dependency MOV-727
Before, the ShutdownFinishedHandler was informed once per registered CapturingListener.
Fortunately, we forced that there was always exactly one listener registered.

Now as STAD wants to unregister and reregister the CapturingListener onPause/onResume,
we had to make sure the Broadcasts to ShutdownFinishedHandler and the STOPPED_ITSELF broadcasts
are always exactly sent once independent of the number of listeners registered.

We also had to add the add/removeListener interfaces to the DataCapturingListener.