[DAT-767] Implement new data transfer protocol (#2)
* Change github secret variable name

* Upgrade dependencies

* Fix build by using Apache Validate method

* [DAT-764] Add protobuf format

* Remove support for truststore (not compatible with google api lib)

* [DAT-764] Add new protocol implementation

* Fix manual SyncPerformerTests

* Fix androidTests

* Cleanup code, add licence texts and increase code versions

* Update Android Build Tools to 7.0.1

This also requires a rather big update of all the build scripts, since Android Build Tools 7.0.1 require Gradle 7+ and this was not compatible with our old build files. Most importantly the way publication was handled was not compatible.

* Ignore DatabaseHelperTest on Apple Silicon

These tests do currently not work on Apple Silicon due to some problems with the SQLIte Library used by Robolectric. They are working on it, but until then we need to ignore the tests on Systems based on the Apple M1/Silicon Architecture.

* Update Protobuf to 3.17.3

Makes it compatible with Apple Silicon/M1

* Update Robolectric to 4.6.1

Does not solve the problem of no compatibility with Apple Silicon/M1 but is a step in that direction, to avoid a large jump as soon as that update is available.

* Use Java 11

Since we upgraded to Android Build Tools 7 it is now possible to use Java 11 for coding. This commit sets this in the build script.

* Fail hard if format is newer then currently supported format

* Simplify protobuf gradle tasks

* Change interface of markAsSynchronized

* Simplify code as suggested

* Use Java 11 in Github Action Workflows

* [DAT-782] Load publish-conventions from android-publish

* Switch back to Cyface Validate method

* Fix Github Workflows

* Upgrade gradle to 7.2

* Fix publish workflow

* Only download proto artifact if not existent

* Replace hardcoded proto version

* Mark measurements deprecated in DataCapturingService constructor

Co-authored-by: Klemens Muthmann <klemens.muthmann@gmail.com>